Added parsing of line number for multicore fitting
Can now parse in a line number which extracts a galaxy list from a specified file that contains a list of galaxies. This can be used to keep all the galaxies to be fit in a particular run within one file, where a line break delineates the separate lists of galaxies.