Updated Quality Control file
Added 67 galaxies which previously could not be modelled because there were issues with cropping the images.
Specifically, Astropy Cutout2D( ) assumes automatically that the x-axis is the right ascension and the y-axis is the declination. This is true for the majority of the images in SDSS except for these in particular.