Deleted FitGalaxies_Server.R
FitGalaxies_Server.R was used for use on a server where a file and line number was specified as the input galaxy list. 
The main FitGalaxies.R script has now been updated to include this galaxy list input feature and so it is superfluous to have two distinct scripts which seemingly perform the same task but under different means of obtaining the inputs. This required some flags which recognised when using the script on a server and dealing with command line arguments that could handle an input file and a reference line number. 

As part of this change, I also removed the append_output( ) function which would output the current optimisation result to a file which contained all the results for a particular optimisation "form". The 'ethos' is now that the results for each optimisation are kept in an individual (appropriately named) file; it is then up to the user to compile these results together afterwards.