Added ability to inherit previous optimisation results
By using the inheritFrom and inheritParams flags in the configuration file, the user may now select a previous optimisation run as well as the specific parameters in which to inherit the initial guesses from. If the no. components in the previous optimisation match that of the new one then the chosen parameters are simply ported across. If however, the parameters to be given to a double-component model come from a single-component fit, various calculations are performed to logically split the parameters fairly between the two components.

** Warning: these features are currently untested **