Fixed bug if previous .RData does not exist
If the previous .RData file that the user was attempting to inherit from did not exist, then the program would still attempt to load it and crash. A flag has been added to catch this in which case the optimisation parameters are simply not inherited.

Also fixed a random 's' that found itself in the code and crashed the program.