Added loadSegMap & loadSigma
loadSegMap and loadSigma can both be set to TRUE in order to load segmentation maps and sigma maps from a fits file located on disk. The files are assumed to be located in the galDir with the filenames:
./[galName]_[band]_SegMap.fits
./[galName]_[band]_Sigma.fits

Future: Check whether the segmentation/sigma maps are the correct dimensions as the image.