Added Save inputs control
The PSF, segmentation map and sigma map that is produced in the program can be saved to the same directory as where the image was taken from by setting the save controls to TRUE in the config file.

Saving inputs will only proceed if load[*] = FALSE (where [*] = PSF|SegMap|Sigma) in the configuration file.