Fixed bug when loading in segmentation maps
When loading in segmentation maps, the segID of a segment does not necessarily correspond the the segstats index (i.e. it's position in the dataframe). Therefore this caused catastrophic errors when defining a segmentation map from a loaded fits file as the segment IDs reference the wrong segments.

To fix this, there are now separate functions for getting the mainID and mainIndex and they are no longer used interchangably throghout the code.