Fixed bug with freeDisk
Fixed a bug when setting freeDisk=FALSE where if the user specified to improve initial guesses, the isophotal fitting would update and add the disk Sersic index as an additional parameter where it should have remained unchanged. Now, the Sersic index of the disk will only update during isophotal fitting if the user allows the disk n to be fit AND they have not taken the Sersic index from a previous fit.