Allow user to load seg map from a different band
The user can now load a segmentation map from a different band than the one being fit using the segMapBand flag in the configuration file. This is only applicable where the loadSegMap has been set to TRUE. Potentially, in the future the segMapBand can also be used to specify that the user wishes to load a different image band and run source finding in that band.