Fixed the if ("stretchscale == auto") bug
An error would occur if the segmentation map was defined over a region of the image where NaNs were present; for example when transferring a segmentation map across images in different bands with offsets in the frame edges. This bug should be fixed now that a line has been added to clean up the segmentation map around areas that contain NaNs in the image.